Extract root folder label disambiguation into a tested helper
Move the leaf-disambiguation algorithm out of `InstanceRootFolder.menuLabel`
into a pure, Foundation-only `RootFolderLabel` namespace so it can be linked
into the host-less Tests target (which has no `@testable import`), mirroring
how `ResolverRouting` backs `InstanceResolver`.

`menuLabel(among:)` now delegates to `RootFolderLabel.disambiguate`. Adds
`RootFolderLabelTests` proving the two layouts from #720: mounts sharing a
leaf directory grow toward the root until unique, while mounts with a
distinct leaf stay a single element.
